collisionSystem fixed, velocity fixed, running

+++ src/Game/cmp/physics.hpp
@@ -9,6 +9,6 @@ struct PhysicsComponent_t : public ECS::ComponentBase_t<PhysicsComponent_t>{
         : ComponentBase_t(eid)
     {};
     
-    uint32_t posX { 0 }, posY { 0 };
-    int32_t velX { 0 }, velY { 0 }; 
+    float posX { 0.0f }, posY { 0.0f };
+    float velX { 0.0f }, velY { 0.0f }; 
 };
